+++ ts/packages/knowledgeProcessor/src/conversation/knowledgeTermSearchSchema2.ts
@@ -23,14 +23,19 @@ export type VerbsTerm = {
     words: string[]; // individual words in single or compound verb
     verbTense: "past" | "present" | "future";
 };
+
+export type SubjectTerm = {
+    subject: string;
+    isPronoun: boolean;
+};
+
 // Action Terms:
 // - when user is querying an action
 // - verb, subject, object and indirectObject associated with the verb
 export type ActionTerm = {
     verbs?: VerbsTerm | undefined; // action verbs
-    // - subject and object for the verb
-    // - exclude all pronouns (e.g. we", "they")
-    subject: string | "none"; // The name of the entity that 'performs' the action. (e.g email sender). Don't include pronouns lke 'they', 'we'..
+    // The name of the entity that 'performs' the action. (e.g. email sender).
+    subject: SubjectTerm | "none";
     object?: string | undefined; // 'receives' the action (e.g. 'email' in: What did X say in his email about Y the sent to Z')
 };
