+++ ts/packages/knowledgeProcessor/src/conversation/searchProcessor.ts
@@ -38,7 +38,6 @@ import {
     TermFilterV2,
 } from "./knowledgeTermSearchSchema2.js";
 import { createTopicSearchOptions } from "./topics.js";
-import { NoEntityName } from "./knowledge.js";
 
 export type SearchProcessingOptions = {
     maxMatches: number;
@@ -573,28 +572,4 @@ export function createSearchProcessor(
         };
         return searchOptions;
     }
-}
-
-export function getAllTermsInFilter(
-    filter: TermFilterV2,
-    includeVerbs: boolean = true,
-): string[] {
-    const action = filter.action;
-    if (action) {
-        let terms: string[] = [];
-        if (includeVerbs && action.verbs) {
-            terms.push(...action.verbs.words);
-        }
-        if (action.subject && action.subject !== NoEntityName) {
-            terms.push(action.subject);
-        }
-        if (action.object) {
-            terms.push(action.object);
-        }
-        if (filter.searchTerms && filter.searchTerms.length > 0) {
-            terms.push(...filter.searchTerms);
-        }
-        return terms;
-    }
-    return filter.searchTerms ?? [];
 }
